Put the connect() call in the same file as the Counter componen and inject redux store at the root
