+++ app/screens/App/screens/Hello/components/Counter.js
@@ -1,13 +1,14 @@
 import React, { Component, PropTypes } from 'react'
+import { connect } from 'react-redux'
 
 class Counter extends Component {
     render() {
-        //const { increment, value } = this.props
+        const { increment, value } = this.props
         return (
             <p>
-                Clicked: {this.props.value} times
+                Clicked: {value} times
                 {' '}
-                <button onClick={this.props.increment}>+</button>
+                <button onClick={increment}>+</button>
             </p>
         )
     }
@@ -18,4 +19,18 @@ Counter.propTypes = {
     value: PropTypes.number.isRequired
 }
 
-export default Counter
+// Which part of the Redux global state does our component want to receive as props?
+function mapStateToProps(state) {
+    return {
+        value: state
+    }
+}
+
+// Which action creators does it want to receive by props?
+function mapDispatchToProps(dispatch) {
+    return {
+        increment: () => dispatch({ type: 'INCREMENT_COUNTER' })
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Counter)
